♻️ refactor : 엔드포인트 및 DTO 수정

+++ src/main/java/com/zip/community/platform/application/service/board/BoardTempService.java
@@ -76,7 +76,12 @@ public class BoardTempService implements TempBoardUseCase {
     }
 
     @Override
-    public void deleteTempBoard(Long boardId) {
+    public void deleteTempBoard(Long boardId, Long userId) {
+
+        if (!tempPort.loadWriterIdByBoardId(boardId).equals(userId)) {
+            throw new CustomException(BoardErrorCode.BAD_REMOVE_BOARD);
+        }
+
         tempPort.deleteTempBoard(boardId);
 
     }
